Only allow logged in user to edit and delete their own added trail

+++ routes/trailroutes.js
@@ -10,6 +10,17 @@ const { isLoggedIn } = require('../middleware');
 const difficulty = ['Easy', 'Moderate', 'Strenuous', 'Extremely strenuous']
 
 
+const isAuthor = async (req, res, next) => {
+    const { id } = req.params;
+    const trail = await Trail.findById(id);
+    if (!trail.author.equals(req.user._id)) {
+        req.flash("error", 'You do not have permission to do that')
+        return res.redirect(`/trails/${id}`);
+    }
+    next();
+}
+
+
 const validateTrail = (req, res, next) => {
     const { error } = trailSchema.validate(req.body);
     if (error) {
@@ -47,7 +58,7 @@ router.get('/:id', catchAsync(async (req, res) => {
     res.render('trails/detail', { trail })
 }))
 
-router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res,) => {
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async (req, res,) => {
     const trail = await Trail.findById(req.params.id);
     if (!trail) {
         req.flash('error', 'Trail not found')
@@ -56,15 +67,20 @@ router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res,) => {
     res.render('trails/edit', { trail, difficulty })
 }))
 
-router.put('/:id', isLoggedIn, validateTrail, catchAsync(async (req, res) => {
+router.put('/:id', isLoggedIn, isAuthor, validateTrail, catchAsync(async (req, res) => {
     const { id } = req.params
-    const trail = await Trail.findByIdAndUpdate(id, { ...req.body })
+    const trailz = await Trail.findByIdAndUpdate(id, { ...req.body })
     req.flash('success', "Successfully updated trail")
-    res.redirect(`/trails/${trail._id}`)
+    res.redirect(`/trails/${trailz._id}`)
 }))
 
 router.delete('/:id', async (req, res) => {
     const { id } = req.params;
+    const trail = await Trail.findById(id);
+    if (!trail.author.equals(req.user._id)) {
+        req.flash("error", 'You do not have permission to do that')
+        return res.redirect(`/trails/${id}`);
+    }
     await Trail.findByIdAndDelete(id);
     req.flash('success', "Deleted trail")
     res.redirect('/trails');
